refactor(chat, renderer): move files
