+++ packages/frameworks/vue/src/chat/ConfigProvider.vue
