+++ packages/frameworks/vue/src/chat/CustomModelProvider.ts
@@ -4,9 +4,9 @@ import { reactive, toRaw } from 'vue';
 import type { LLMConfig, IGenuiConfig, ICustomConfig } from './chat.types';
 import { emitter } from './event-emitter';
 import useSchemaStream from './useSchemaStream';
-import type { IStreamDelta, IMessageItem, IChatMessage } from '../../core/chat';
+import type { IStreamDelta, IMessageItem, IChatMessage } from '../../../../core/src/protocols/chat'; //TODO: replace with package name
 import { v4 as uuidv4 } from 'uuid';
-import { useI18n } from '../i18n';
+import { useI18n } from './i18n';
 
 export interface ICustomModelProviderOptions {
   url: string;
