+++ packages/frameworks/vue/src/chat/chat-utils.ts
