+++ packages/frameworks/vue/src/chat/composable/use-resize.ts
