+++ packages/frameworks/vue/src/chat/continue-chat-action.ts
@@ -1,4 +1,4 @@
-import { useI18n } from '../i18n';
+import { useI18n } from './i18n';
 
 export function useContinueChatAction(
   onAction: (_: { llmFriendlyMessage: string; humanFriendlyMessage: string; context: Record<string, any> }) => void,
