+++ packages/frameworks/vue/src/chat/event-emitter.ts
