+++ packages/frameworks/vue/src/chat/file-upload/file-utils.ts
