+++ packages/frameworks/vue/src/chat/i18n/en.json
