+++ packages/frameworks/vue/src/chat/i18n/i18n.ts
