+++ packages/frameworks/vue/src/chat/i18n/index.ts
