+++ packages/frameworks/vue/src/chat/i18n/zh.json
