+++ packages/frameworks/vue/src/chat/renderer/AttachmentsRenderer.vue
