+++ packages/frameworks/vue/src/chat/renderer/TemplateDataRenderer.vue
