+++ packages/frameworks/vue/src/chat/renderer/ToolRenderer.vue
