+++ packages/frameworks/vue/src/chat/useSchemaStream.ts
