+++ packages/genui-renderer/src/vue/index.ts
@@ -1,8 +1,8 @@
 export { default as SchemaRenderer } from '../../../frameworks/vue/src/renderer/SchemaCardRenderer.vue'; //TODO: replace with package name
-export { default as ConfigProvider } from './ConfigProvider.vue';
-export { default as GenuiChat } from './genui-chat/index.vue';
+export { default as ConfigProvider } from '../../../frameworks/vue/src/chat/ConfigProvider.vue'; //TODO: replace with package name
+export { default as GenuiChat } from '../../../frameworks/vue/src/chat/index.vue';
 export { RENDERER_SETTINGS_KEY } from '@opentiny/tiny-schema-renderer';
-export * from './genui-chat/chat.types';
+export * from '../../../frameworks/vue/src/chat/chat.types.js';
 export * from '../../../frameworks/vue/src/renderer/renderer.types'; //TODO: replace with package name
-export { SCHEMA_RENDERER_INJECTION_TOKEN } from './injection-tokens.js';
-export * from './i18n';
+export { SCHEMA_RENDERER_INJECTION_TOKEN } from '../../../frameworks/vue/src/chat/injection-tokens.js';
+export * from '../../../frameworks/vue/src/chat/i18n'; //TODO: replace with package name
